Cleanup pkg/cgroups for compilation on freebsd
freebsd and other non linux systems do not support
dbus.
