+++ pkg/cgroups/cgroups_unsupported.go
@@ -3,10 +3,7 @@
 package cgroups
 
 import (
-	"fmt"
 	"os"
-
-	systemdDbus "github.com/coreos/go-systemd/v22/dbus"
 )
 
 // IsCgroup2UnifiedMode returns whether we are running in cgroup 2 cgroup2 mode.
@@ -22,9 +19,4 @@ func UserOwnsCurrentSystemdCgroup() (bool, error) {
 
 func rmDirRecursively(path string) error {
 	return os.RemoveAll(path)
-}
-
-// UserConnection returns an user connection to D-BUS
-func UserConnection(uid int) (*systemdDbus.Conn, error) {
-	return nil, fmt.Errorf("systemd d-bus is not supported on this platform")
 }
